dj_msqrvve_brand_system/requirements.txt
requests
PyYAML
pytest
python-dotenv
responses
Flask
pkce
